refactor(kernel/testsuite): common functions into `testsuite/common`
As seen on https://doc.rust-lang.org/rust-by-example/testing/integration_testing.html
